Move BoardDefaults to the library
Also use generic names to identify Boards.
