+++ grovebutton/build.gradle
@@ -31,5 +31,6 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile project(':driverlibrary')
     compile 'io.mraa.at.upm:upm_grove:1.1.0'
 }
