+++ groverelay/src/main/java/com/example/upm/androidthings/driversamples/GroveRelay.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2017 Intel Corporation.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.example.upm.androidthings.driversamples;
 
 import android.app.Activity;
@@ -5,7 +21,9 @@ import android.os.Bundle;
 import android.os.Build;
 import android.util.Log;
 
+import com.example.upm.androidthings.driverlibrary.BoardDefaults;
 import mraa.mraa;
+
 public class GroveRelay extends Activity {
 
     private static final String TAG = "GroveRelayActivity";
