+++ mma7660/src/main/java/com/example/upm/androidthings/driversamples/AccelerometerActivity.java
@@ -26,6 +26,7 @@ import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 
 import com.example.upm.androidthings.driverlibrary.Mma7660AccelerometerDriver;
+import com.example.upm.androidthings.driverlibrary.BoardDefaults;
 import mraa.mraa;
 
 import java.io.IOException;
@@ -50,13 +51,13 @@ public class AccelerometerActivity extends Activity implements SensorEventListen
 
         switch (bd.getBoardVariant()) {
             case BoardDefaults.DEVICE_EDISON_ARDUINO:
-                i2cIndex = mraa.getI2cLookup(getString(R.string.ACCEL_Edison_Arduino));
+                i2cIndex = mraa.getI2cLookup(getString(R.string.Edison_Arduino));
                 break;
             case BoardDefaults.DEVICE_EDISON_SPARKFUN:
-                i2cIndex = mraa.getI2cLookup(getString(R.string.ACCEL_Edison_Sparkfun));
+                i2cIndex = mraa.getI2cLookup(getString(R.string.Edison_Sparkfun));
                 break;
             case BoardDefaults.DEVICE_JOULE_TUCHUCK:
-                i2cIndex = mraa.getI2cLookup(getString(R.string.ACCEL_Joule_Tuchuck));
+                i2cIndex = mraa.getI2cLookup(getString(R.string.Joule_Tuchuck));
                 break;
             default:
                 throw new IllegalStateException("Unknown Board Variant: " + bd.getBoardVariant());
